Lock Screen - Add Lockscreen Wallpaper (1/2)
PS3:
When the status bar uses a transparent theme the standard
wallpaper can be seen behind it and this is visually
unapealing when using a custom lockscreen wallpaper.  This
patch removes the FLAG_SHOW_WALLPAPER from the Keyguard so
the statusbar will be opaque.  In the case a lockscreen
wallpaper is not set we will load the default wallpaper
and display that in the ImageView.

Change-Id: Ibdfff3a1756647e2cb88106e4d79d500a271facc